use mithril to render views
