+++ client/index.js
@@ -1,15 +1,73 @@
-const io = window.io
-const id = /\/([^\/]+)$/.exec(location.pathname)[1]
-const socket = io()
+const socket = require('socket.io-client')()
+const Editor = require('./component/editor')
+const m = require('mithril')
+const id = window.id
+const href = location.href
 
-const updatePageURI = require('./updatePageURI')
-const startSyncNote = require('./startSyncNote')
-const startMonitorNetwork = require('./startMonitorNetwork')
+const App = {
+  networkStatus: '',
+  saveStatusClass: '',
+  saveStatusTimer: null,
 
-function start() {
-  updatePageURI({ href: location.href })
-  startSyncNote({ id, socket })
-  startMonitorNetwork({ socket })
+  oninit() {
+    // monitor network
+    const events = {
+      connect: '',
+      reconnect: '',
+      reconnect_attempt: 'connection lost',
+      connect_error: 'connection lost',
+      connect_timeout: 'connection lost',
+      reconnect_error: 'connection lost',
+      reconnect_failed: 'connection lost'
+    }
+    Object.keys(events).forEach(evt =>
+      socket.on(evt, () => {
+        this.networkStatus = events[evt]
+        m.redraw()
+      })
+    )
+  },
+
+  onSaveStatusChange(isSaving) {
+    clearTimeout(this.saveStatusTimer)
+    if (isSaving) {
+      this.saveStatusClass = 'show'
+      m.redraw()
+    } else {
+      this.saveStatusTimer = setTimeout(() => {
+        this.saveStatusClass = ''
+        m.redraw()
+      }, 300)
+    }
+  },
+
+  view() {
+    return m(
+      'main',
+      m('header', [
+        m('small#save-status', { class: this.saveStatusClass }, 'saving'),
+        m('small#network-status', this.networkStatus)
+      ]),
+      m(
+        'section',
+        m(
+          '.layer',
+          m(
+            '.layer',
+            m(
+              '.layer',
+              m(Editor, {
+                socket,
+                id,
+                onStatusChange: this.onSaveStatusChange.bind(this)
+              })
+            )
+          )
+        )
+      ),
+      m('footer', m('small', m('a.this-page', { href }, href)))
+    )
+  }
 }
 
-start()
+m.mount(document.body, App)
