--- client/updatePageURI.js
@@ -1,4 +1,0 @@
-module.exports = function updatePageURI({ href }) {
-  const link = document.getElementById('this-page')
-  link.href = (link.textContent = href)
-}
