+++ client/util/stream.js
