+++ package.json
@@ -11,12 +11,14 @@
     "koa-router": "^7.0.1",
     "koa-static": "^3.0.0",
     "koa-views": "^6.0.1",
+    "mithril": "^1.1.3",
     "most": "^1.2.2",
     "mustache": "^2.2.1",
     "nedb": "^1.8.0",
     "node-diff3": "0.0.1",
     "randomstring": "^1.1.5",
     "socket.io": "^1.4.8",
+    "socket.io-client": "^2.0.3",
     "string-hash": "^1.1.0",
     "webpack": "^2.3.2"
   },
