+++ server/views/app.html
@@ -13,29 +13,13 @@
   <title>{{id}} · notepad</title>
   <link rel="stylesheet" href="./style/main.css">
   <link rel="manifest" href="./manifest.json">
-  <script src='./vendor/socket.io-1.4.5.js' defer></script>
+  <script>
+    window.id = '{{{id}}}'
+  </script>
   <script src='./script/bundle.js' defer></script>
 </head>
 
 <body>
-  <main>
-    <header>
-      <small id="save-status">saving</small>
-      <small id="network-status"></small>
-    </header>
-    <section>
-      <div class="layer">
-        <div class="layer">
-          <div class="layer">
-            <textarea name="editor" id="editor">{{note}}</textarea>
-          </div>
-        </div>
-      </div>
-    </section>
-    <footer>
-      <small><a id='this-page' class='this-page'></a></small>
-    </footer>
-  </main>
 </body>
 <script>
   (function(i, s, o, g, r, a, m) {
